StockNewsAlert/requirements.txt
requests
python-dotenv
pandas
twilio
azure-identity
azure-keyvault-secrets
pyyaml
jinja2
